Create app.rs
